requirements.txt
Flask
groq
python-decouple
pandas
openpyxl
gunicorn
celery
redis
requests